Updating main.dart
fixes: #10
